+++ changelog.md
@@ -1,3 +1,7 @@
+## v1.9
+- Addon list show addon memory usage now
+- Addon tooltip show addon version now
+- Support switch configuration macro, use "/impal help" to learn more 
 ## v1.8.17
 
 - Fix error: delete a global addon loading scheme will cause anothor player which use that scheme error